Refactor remove outliers into separate script
